+++ pkg/providers/appsec/testdata/TestDSWafAIRules/waf_ai_rules.tf
@@ -1,0 +1,9 @@
+provider "akamai" {
+  edgerc        = "../../common/testutils/edgerc"
+  cache_enabled = false
+}
+
+data "akamai_appsec_waf_ai_rules" "test" {
+  config_id          = 111111
+  security_policy_id = "2222_333333"
+}