+++ pkg/providers/appsec/testdata/TestDSWafAIRules/waf_ai_rules_missing_config_id.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc        = "../../common/testutils/edgerc"
+  cache_enabled = false
+}
+
+data "akamai_appsec_waf_ai_rules" "test" {
+  security_policy_id = "2222_333333"
+}